Admin can now post code

+++ lib/discordbot/filters/code.ex
@@ -2,9 +2,10 @@ defmodule Discordbot.Code do
   
   alias DiscordEx.RestClient.Resources.Channel
   alias Discordbot.Helpers.Message
+  alias Discordbot.Commands.Admin
 
   def handle(:message_create, payload, state = %{rest_client: conn}) do
-    if is_code(payload["content"]) do
+    if is_code(payload["content"]) && !Admin.is_admin?(payload) do
       spawn fn -> Channel.delete_message(conn, payload["channel_id"], payload["id"]) end
       spawn fn ->
         case Message.dm(conn, payload["author"]["id"], dm(payload)) do
